ml_core/requirements.txt
pandas
numpy
scikit-learn
xgboost
mlflow
kafka-python
joblib
python-dotenv
